int/uint fixed

+++ Source/Network/Polytempo_OSCListener.cpp
@@ -127,10 +127,10 @@ void Polytempo_OSCListener::oscMessageReceived(const OSCMessage & message)
         }
         
         // calculate syncTime
-        int syncTime;
+        uint32 syncTime;
         
         if(event->hasProperty(eventPropertyString_TimeTag))
-            syncTime = event->getProperty(eventPropertyString_TimeTag);
+            syncTime = uint32(int(event->getProperty(eventPropertyString_TimeTag)));
 		else
 		{
 			Polytempo_TimeProvider::getInstance()->getSyncTime(&syncTime);
